app/core/config.py
import os
from dotenv import load_dotenv

load_dotenv()

def get_env(name: str) -> str:
    value = os.getenv(name, "").strip()
    if not value:
        raise RuntimeError(f"Missing environment variable: {name}")
    return value

SHOPIFY_SHOP = get_env("SHOPIFY_SHOP")
SHOPIFY_ACCESS_TOKEN = get_env("SHOPIFY_ACCESS_TOKEN")
SHOPIFY_API_VERSION = os.getenv("SHOPIFY_API_VERSION", "2025-07").strip() or "2025-07"