requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
httpx==0.28.1
jinja2==3.1.4
python-multipart==0.0.12
python-dotenv==1.0.1